% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMB.cv.r
\name{BMB.cv}
\alias{BMB.cv}
\title{BMB critical value for the maximum of Chatterjee's rank correlations}
\usage{
BMB.cv(dat, q, B, alpha, type = "bmb1", seed = NULL)
}
\arguments{
\item{dat}{List with two elements: vector X and matrix Y. The \eqn{j}-th column in \eqn{Y} corresponds to \eqn{Y_j} in the individual hypothesis \eqn{Y_j \perp X}.}

\item{q}{Numeric. The block size.}

\item{B}{Numeric. The number of bootstrap samples.}

\item{alpha}{Numeric. The significance level.}

\item{type}{Character. This argument specifies whether and how the test statistic and the bootstrap statistic are studentized. Options are "bmb" (no studentization), "bmb1" (default option), and "bmb2" (alternative studentization). The types are formally described in Olivares, Olma, and Wilhelm (2025).}

\item{seed}{Numeric. The seed for the random number generator. If \code{NULL}, the seed is not set. If a positive integer, it sets the seed for reproducibility.}
}
\value{
Returns a list with the critical value and the type of test statistic.
}
\description{
This function returns the bootstrap critical value of the statistic \deqn{T^B=\max_{1\le j\le p}\frac{1}{\sqrt{mq}}\sum_{k=1}^{m} \hat{A}_{j,k} \cdot \varepsilon_{k},} where the \eqn{\varepsilon_{k}} are independent standard normal random variables, based on \eqn{B} block multiplier bootstrap samples.
}
\examples{
\dontrun{
n <- 100
p <- 10
dat <- list()
dat$X <- rnorm(n)
dat$Y <- MASS::mvrnorm(n = n, mu = rep(0,  p), Sigma=diag(rep(1, p)))
BMB.cv(dat, q=2, B=100, alpha=0.05, seed = 5) # default type is "bmb1"

}
}
\references{
Olivares, M., Olma, T., and Wilhelm, D. (2025). A Powerful Bootstrap Test of Independence in High Dimensions. Preprint, arXiv:2503.21715.
}
\author{
Maurcio Olivares

Tomasz Olma

Daniel Wilhelm
}
\keyword{Chatterjee}
\keyword{Correlation}
\keyword{Rank}
\keyword{block}
\keyword{bootstrap}
\keyword{mutiplier}
