% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_RomanoWolf.r
\name{stepdown_RomanoWolf}
\alias{stepdown_RomanoWolf}
\title{Romano-Wolf stepdown method}
\usage{
stepdown_RomanoWolf(
  dat,
  q,
  B,
  alpha,
  type = "bmb1",
  seed = NULL,
  steps = FALSE
)
}
\arguments{
\item{dat}{List with two elements: vector X and matrix Y. The \eqn{j}-th column in \eqn{Y} corresponds to \eqn{Y_j} in the individual hypothesis \eqn{Y_j \perp X}.}

\item{q}{Numeric. The block size for the block multiplier bootstrap.}

\item{B}{Numeric. The number of bootstrap replications.}

\item{alpha}{Numeric. The significance level.}

\item{type}{Character. This argument specifies whether and how the test statistic and the bootstrap statistic are studentized. Options are "bmb" (no studentization), "bmb1" (default option), and "bmb2" (alternative studentization). The types are formally described in Olivares, Olma, and Wilhelm (2025).}

\item{seed}{Numeric. The seed for the random number generator. If \code{NULL}, the seed is not set. If a positive integer, it sets the seed for reproducibility.}

\item{steps}{Logical. If \code{TRUE}, the function will return the steps of the stepdown procedure. The default is \code{FALSE}.}
}
\value{
Returns a list with the total number of rejections. If steps = \code{TRUE}, it will also return the steps of the stepdown procedure.
}
\description{
Consider a family of individual null hypotheses \eqn{H_{0,j}: Y_j \perp X}, for \eqn{j=1,\dots,p}. This function implements a stepwise procedure for selecting those variables from \eqn{Y_1,\dots,Y_p} that violate independence, while controlling the family-wise error rate.
}
\references{
Olivares, M., Olma, T., and Wilhelm, D. (2025). A Powerful Bootstrap Test of Independence in High Dimensions. Preprint, arXiv:2503.21715.
}
\author{
Maurcio Olivares

Tomasz Olma

Daniel Wilhelm
}
\keyword{Block}
\keyword{Bootstrap}
\keyword{Chatterjee}
\keyword{Correlation}
\keyword{Multiplier}
\keyword{Rank}
\keyword{Romano-Wolf}
\keyword{Stepdown}
