% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_size.r
\name{block_size}
\alias{block_size}
\title{Blocks size for Block Multiplier Bootstrap.}
\usage{
block_size(n)
}
\arguments{
\item{n}{Numeric. Sample size}
}
\value{
Returns an integer value of block size.
}
\description{
Calculates size of the blocks needed for the block multiplier bootstrap, denoted as \eqn{q}, using the rule of thumb described in Olivares, Olma, and Wilhelm (2025). The function returns \eqn{q} using the approximatation if the sample size is larger than 224.
}
\references{
Olivares, M., Olma, T., and Wilhelm, D. (2025). A Powerful Bootstrap Test of Independence in High Dimensions. Preprint, arXiv:2503.21715.
}
\author{
Maurcio Olivares

Tomasz Olma

Daniel Wilhelm
}
\keyword{Bootstrap}
\keyword{Chatterjee}
\keyword{Correlation}
\keyword{Rank}
