% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMB_indep.r
\name{BMB_indep}
\alias{BMB_indep}
\title{Block-multiplier bootstrap test of independence}
\usage{
BMB_indep(dat, B, alpha, type = "bmb1", seed = NULL)
}
\arguments{
\item{dat}{List with two elements: vector X and matrix Y. The \eqn{j}-th column in \eqn{Y} corresponds to \eqn{Y_j} in the individual hypothesis \eqn{Y_j \perp X}.}

\item{B}{Numeric. The number of bootstrap samples.}

\item{alpha}{Numeric. The significance level.}

\item{type}{Character. This argument specifies whether and how the test statistic and the bootstrap statistic are studentized. Options are "bmb" (no studentization), "bmb1" (default option), and "bmb2" (alternative studentization). The types are formally described in Olivares, Olma, and Wilhelm (2025).}

\item{seed}{Numeric. The seed for the random number generator. If \code{NULL}, the seed is not set. If a positive integer, it sets the seed for reproducibility.}
}
\value{
An object of class "BMB_indep", a list containing the following components:
\item{type}{Type. See description above}
\item{T_obs}{The value of the test statistic.}
\item{n}{Sample Size.}
\item{p}{Number of hypotheses.}
\item{block_size}{Block size used in the bootstrap.}
\item{B}{Number of bootstrap samples.}
\item{cv}{Bootstrap critical value corresponding to the test statistic.}
}
\description{
This function tests the hypothesis \eqn{H_{0}: Y_j \perp X} for \eqn{j=1,\dots,p} using the block multiplier bootstrap test of Olivares, Olma, and Wilhelm (2025).
}
\examples{
\dontrun{
n <- 100
p <- 10
dat <- list()
dat$X <- rnorm(n)
dat$Y <- MASS::mvrnorm(n = n, mu = rep(0,  p), Sigma=diag(rep(1, p)))
test_indep <- BMB_indep(dat, B=100, alpha=0.05, type = "bmb", seed = 5)
summary(test_indep)

}
}
\references{
Olivares, M., Olma, T., and Wilhelm, D. (2025). A Powerful Bootstrap Test of Independence in High Dimensions. Preprint, arXiv:2503.21715.
}
\author{
Maurcio Olivares

Tomasz Olma

Daniel Wilhelm
}
\keyword{Chatterjee}
\keyword{Correlation}
\keyword{Rank}
