% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BMB_indep.r
\name{summary.BMB_indep}
\alias{summary.BMB_indep}
\title{Summarizing Block Multiplier Bootstrap Test of Independence in High Dimensions.}
\usage{
\method{summary}{BMB_indep}(object, ..., digits = 4)
}
\arguments{
\item{object}{an object of class \code{"BMB_indep"}, usually a result of a call to \code{\link{BMB_indep}}.}

\item{...}{unused}

\item{digits}{number of digits to display}
}
\value{
\code{summary_BMB_indep} returns an object of \link{class} "\code{summary.BMB_indep}" which has the following components
 \item{results}{Matrix with the Testing Problem, Test Statistics, type specifying whether the test statistic and the bootstrap statistic are studentized, Sample Size, Number of Hypothesis, Significance Level, Block size, Number of bootrap samples, and Block-Multiplier-Bootstrap critical value.}
}
\description{
\code{summary} method for class \code{"BMB_indep"}
}
\author{
Maurcio Olivares

Tomasz Olma

Daniel Wilhelm
}
